CRUD completed with authentication

+++ backend/controllers/goalControllers.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler');
 
-const Goal=require('../config/model/goalModel')
+const Goal=require('../config/model/goalModel');
+const User = require('../config/model/userModel');
 
 
 //@desc    Get goal: 
 //@route   GET /api/goals
 //@access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals= await Goal.find()
+    const goals= await Goal.find({user: req.user.id})
     res.status(200).json(goals);
 })
 
@@ -21,7 +22,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field ')
     }
     const goal= await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -35,6 +37,17 @@ const updateGoal =asyncHandler( async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user=await User.findById(req.user.id)
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+    //ensures loged in user matches goal user
+    if(goal.user.toString()!==user.id){
+        res.status(401)
+        throw new Error('User not authorised')
     }
     const updatedGoal= await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true,})
 
@@ -50,6 +63,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user=await User.findById(req.user.id)
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
+    }
+    //ensures loged in user matches goal user
+    if(goal.user.toString()!==user.id){
+        res.status(401)
+        throw new Error('User not authorised')
     }
     await goal.deleteOne();     
     res.status(200).json({id:req.params.id});
